Map from source image ID to internal ID
The Dataset class assigns int image IDs to all
images, but keeps track of their original ID from
the source dataset. It’s easy to get the source ID
if you have the image_id. This new mapping makes
it also easy to get the image_id if you have the
source id.